requirements.txt
flask
gpiozero
pyyaml
